#5 сделала классы в отдельных файлах и консоль
