+++ .vs/Study/xs/UserPrefs.xml
@@ -2,7 +2,11 @@
   <MonoDevelop.Ide.ItemProperties.Study PreferredExecutionTarget="MonoDevelop.Default" />
   <MonoDevelop.Ide.Workbench ActiveDocument="Study/Program.cs">
     <Files>
-      <File FileName="Study/Program.cs" Line="44" Column="43" />
+      <File FileName="Study/Program.cs" Line="134" Column="5" />
+      <File FileName="Study/Forecast.cs" Line="60" Column="10" />
+      <File FileName="Study/Person.cs" Line="21" Column="32" />
+      <File FileName="Study/User.cs" Line="4" Column="31" />
+      <File FileName="Study/Admin.cs" Line="1" Column="1" />
     </Files>
     <Pads>
       <Pad Id="ProjectPad">
@@ -13,6 +17,9 @@
             </Node>
           </Node>
         </State>
+      </Pad>
+      <Pad Id="MonoDevelop.Debugger.WatchPad">
+        <State />
       </Pad>
     </Pads>
   </MonoDevelop.Ide.Workbench>
